Make a cube class that allows the user to draw edges.
Previously, the cube was defined as a set of four faces. While this
representation worked, what you saw on screen was just the 8 vertices of
the cube. This update makes it so that is possible to have lines
connecting the vertices. Since ncurses cannot draw arbitrary lines, we
must calculate the points in 3D space that form the edges of the cube.
The new Cube class adds optional support for computing additional edge
points on the cube.